feat(style-modify): extract style-setting-section
